# PERF-NETWORK-TCP-SINGLE-CONNECTION-THROUGHPUT-SYNTHETIC - add results in test summary 