Before:
```
[LISAv2 Test Results Summary]
Test Run On           : 05/24/2019 06:34:19
ARM Image Under Test  : canonical : ubuntuserver : 18.04-lts : Latest
Total Test Cases      : 1 (1 Passed, 0 Failed, 0 Aborted, 0 Skipped)
Total Time (dd:hh:mm) : 0:0:49

   ID TestArea             TestCaseName                                                                TestResult TestDuration(in minutes) 
-------------------------------------------------------------------------------------------------------------------------------------------
    1 NETWORK              PERF-NETWORK-TCP-SINGLE-CONNECTION-THROUGHPUT-SYNTHETIC                           PASS                45.74 
```


After:
```
[LISAv2 Test Results Summary]
Test Run On           : 05/24/2019 06:34:19
ARM Image Under Test  : canonical : ubuntuserver : 18.04-lts : Latest
Total Test Cases      : 1 (1 Passed, 0 Failed, 0 Aborted, 0 Skipped)
Total Time (dd:hh:mm) : 0:0:49

   ID TestArea             TestCaseName                                                                TestResult TestDuration(in minutes) 
-------------------------------------------------------------------------------------------------------------------------------------------
    1 NETWORK              PERF-NETWORK-TCP-SINGLE-CONNECTION-THROUGHPUT-SYNTHETIC                           PASS                45.74 
	BufferLengths=32 : tx_throughput=0.11Gbps rx_throughput=0.11Gbps retransmitted_segments=0 congestion_windowsize_kb=139 
	BufferLengths=64 : tx_throughput=0.21Gbps rx_throughput=0.21Gbps retransmitted_segments=0 congestion_windowsize_kb=288 
	BufferLengths=128 : tx_throughput=0.41Gbps rx_throughput=0.41Gbps retransmitted_segments=0 congestion_windowsize_kb=502 
	BufferLengths=256 : tx_throughput=0.80Gbps rx_throughput=0.80Gbps retransmitted_segments=0 congestion_windowsize_kb=1222 
	BufferLengths=512 : tx_throughput=1.54Gbps rx_throughput=1.54Gbps retransmitted_segments=0 congestion_windowsize_kb=2269 
	BufferLengths=1024 : tx_throughput=3.14Gbps rx_throughput=3.14Gbps retransmitted_segments=1 congestion_windowsize_kb=2758 
	BufferLengths=2048 : tx_throughput=3.7Gbps rx_throughput=3.7Gbps retransmitted_segments=8 congestion_windowsize_kb=2831 
	BufferLengths=4096 : tx_throughput=3.78Gbps rx_throughput=3.78Gbps retransmitted_segments=0 congestion_windowsize_kb=2719 
	BufferLengths=8192 : tx_throughput=3.66Gbps rx_throughput=3.66Gbps retransmitted_segments=72 congestion_windowsize_kb=2856 
	BufferLengths=16384 : tx_throughput=3.82Gbps rx_throughput=3.81Gbps retransmitted_segments=36 congestion_windowsize_kb=2654 
	BufferLengths=32768 : tx_throughput=3.82Gbps rx_throughput=3.82Gbps retransmitted_segments=158 congestion_windowsize_kb=2432 
	BufferLengths=65536 : tx_throughput=3.85Gbps rx_throughput=3.85Gbps retransmitted_segments=205 congestion_windowsize_kb=2514 
```